move cola minimal pair out of tasks.py
