+++ src/tasks/__init__.py
@@ -5,6 +5,7 @@ from . import lm
 from . import mt
 from . import nli_probing
 from . import reddit
+from . import minimal_pairs
 
 # REGISTRY needs to be available to modules within this package,
 # but we also import it here to make it available at the package level.
